Update readMe
This mobile xamarin app is called bloodApp, Users can sign in and signUp, after signing in, user will have 4 option on the HomePage, the first option is Find BloodGroup, it will take him to another page and now he can choose his blood group and his country and then he can search for all users according to the blood group and the country, supposing we want a blood donation from this user, we click on the user and we will get his blood group, image profile of this user country, we also can call this user or send him and email.  the second option is Register blood group, this option means if we are a donar and we want to doante blood we press the register button and then we need to provide our information like: Name, Email, Phone, Bloodgroup , country and we need to upload our profile picture, we will open the device Camera and we will take a photo and save it and finally we will submit our profile. The third option is the latest donars, it means we will get all the laters Donars.